feat: highlight duplicated queries in the UI and improve styling
