+++ src/renderer/components/navbar/NavBarSettings.vue
@@ -1,22 +1,12 @@
 <script setup>
-import { onMounted, ref } from "vue";
+import { ref } from "vue";
 import { CogIcon, HomeIcon } from "@heroicons/vue/24/outline";
-import router from "../../router";
-import { useXDebug } from "@/store/xdebug";
-
-const xDebugStore = useXDebug();
 
 const inSettingPage = ref(false);
 
 const togglePage = () => {
     inSettingPage.value = !inSettingPage.value;
 };
-
-onMounted(() => {
-    window.ipcRenderer.on("new.dumps", (event) => {
-        router.push({ name: "home" }, { xdebug: xDebugStore.current !== "" });
-    });
-});
 </script>
 
 <template>
